refactor: Image, Icon 컨벤션 수정
