+++ src/assets/icons/index.ts
@@ -1,3 +1,3 @@
-export { ReactComponent as Alert } from "./alert.svg";
-export { ReactComponent as IcCool } from "./coolicon.svg";
-export { ReactComponent as IcLogo } from "./logoicon.svg";
+export { ReactComponent as IcAlert } from "./alertIcon.svg";
+export { ReactComponent as IcCool } from "./coolIcon.svg";
+export { ReactComponent as IcLogo } from "./logoIcon.svg";
