+++ src/components/signup/ThirdStep.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { Alert } from "../../assets/icons";
+import { IcAlert } from "../../assets/icons";
 
 export default function ThirdStep() {
   return (
@@ -21,12 +21,12 @@ export default function ThirdStep() {
         <label></label>
         <input type="text" id="email" placeholder="영문, 숫자, 특수문자를 조합해 8자 이상 입력해 주세요" />
         <label>
-          <Alert />
+          <IcAlert />
           <strong>비밀번호 형식 에러</strong>
         </label>
         <input type="text" id="email" placeholder="비밀번호를 확인해 주세요" />
         <label>
-          <Alert />
+          <IcAlert />
           <strong>비밀번호가 다릅니다.</strong>
         </label>
         <button>다음 계단</button>
